# docs/openapi.yaml
openapi: 3.0.3
info:
  title: XS API
  version: 0.1.0
servers:
  - url: http://localhost:3001/api
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ApiOk:
      type: object
      required: [success, data]
      properties:
        success:
          type: boolean
          enum: [true]
        data: {}
        traceId:
          type: string
          nullable: true
    ApiErr:
      type: object
      required: [success, error]
      properties:
        success:
          type: boolean
          enum: [false]
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
        traceId:
          type: string
          nullable: true
    Permission:
      type: object
      required: [id, code]
      properties:
        id: { type: string }
        code: { type: string }
        description: { type: string, nullable: true }
    RoleListItem:
      type: object
      required: [id, name, _count]
      properties:
        id: { type: string }
        name: { type: string }
        description: { type: string, nullable: true }
        _count:
          type: object
          required: [users, permissions]
          properties:
            users: { type: integer }
            permissions: { type: integer }
    RoleDetail:
      type: object
      required: [id, name, permissions, _count]
      properties:
        id: { type: string }
        name: { type: string }
        description: { type: string, nullable: true }
        permissions:
          type: array
          items:
            type: object
            required: [permission]
            properties:
              permission:
                $ref: '#/components/schemas/Permission'
        _count:
          type: object
          required: [users]
          properties:
            users: { type: integer }
paths:
  /auth/login:
    post:
      security: []
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string }
                password: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [user, token]
                properties:
                  user:
                    type: object
                  token:
                    type: string
  /roles:
    get:
      summary: List roles (paginated)
      parameters:
        - in: query
          name: page
          schema: { type: integer, minimum: 1, default: 1 }
        - in: query
          name: pageSize
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
        - in: query
          name: q
          schema: { type: string }
      responses:
        '200':
          description: OK
          headers:
            x-trace-id:
              schema: { type: string }
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiOk'
                  - type: object
                    properties:
                      data:
                        type: object
                        required: [items, total, page, pageSize]
                        properties:
                          items:
                            type: array
                            items: { $ref: '#/components/schemas/RoleListItem' }
                          total: { type: integer }
                          page: { type: integer }
                          pageSize: { type: integer }
        '401':
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiErr' }
        '403':
          description: Forbidden
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiErr' }
    post:
      summary: Create role
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
                description: { type: string }
                permissionIds:
                  type: array
                  items: { type: string }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiOk'
                  - type: object
                    properties:
                      data:
                        type: object
                        required: [id, name]
                        properties:
                          id: { type: string }
                          name: { type: string }
                          description: { type: string, nullable: true }
        '409':
          description: Conflict
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiErr' }
  /roles/permissions:
    get:
      summary: List permissions (cached)
      parameters:
        - in: query
          name: q
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiOk'
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/Permission' }
  /roles/export:
    get:
      summary: Export roles as CSV
      parameters:
        - in: query
          name: format
          schema:
            type: string
            enum: [csv]
            default: csv
      responses:
        '200':
          description: CSV
          content:
            text/csv:
              schema:
                type: string
  /roles/bulk-delete:
    post:
      summary: Bulk delete roles
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ids]
              properties:
                ids:
                  type: array
                  items: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiOk'
                  - type: object
                    properties:
                      data:
                        type: object
                        required: [deleted]
                        properties:
                          deleted: { type: integer }
  /roles/{id}:
    get:
      summary: Role detail
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiOk'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/RoleDetail' }
    put:
      summary: Update role basic info
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                description: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiOk' }
    delete:
      summary: Delete role
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiOk' }
        '409':
          description: Role in use
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiErr' }
  /roles/{id}/permissions:
    put:
      summary: Replace role permissions
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [permissionIds]
              properties:
                permissionIds:
                  type: array
                  items: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiOk' }